move kmip-role-fields to model-helpers
